# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Api Documentation
  description: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Inferred Url
...
paths:
...
  /appointment/delete/{appointmentId}:
    delete:
      tags:
        - appointment-controller
      summary: Deletes an appointment.
      operationId: deleteAppointmentUsingDELETE
      parameters:
        - name: appointmentId
          in: path
          description: appointmentId
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyBody'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /appointment/new:
    post:
      tags:
        - appointment-controller
      summary: Creates a new appointment.
      operationId: getAppointmentUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRequest'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /appointment/update:
    put:
      tags:
        - appointment-controller
      summary: Updates an appointment.
      operationId: updateAppointmentUsingPUT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRequest'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /appointment/{appointmentId}:
    get:
      tags:
        - appointment-controller
      summary: Search appointment by ID.
      operationId: getAppointmentUsingGET
      parameters:
        - name: appointmentId
          in: path
          description: appointmentId
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRes'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /appointment/{appointmentId}/status/{status}:
    put:
      tags:
        - appointment-controller
      summary: Updates the status of an appointment by ID.
      operationId: updateStatusUsingPUT
      parameters:
        - name: appointmentId
          in: path
          description: appointmentId
          required: true
          style: simple
          schema:
            type: integer
            format: int64
        - name: status
          in: path
          description: status
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRes'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
components:
  schemas:
...
    AppointmentReq:
      title: AppointmentReq
      type: object
      properties:
        barber:
          $ref: '#/components/schemas/BarberReq'
        customer:
          $ref: '#/components/schemas/CustomerReq'
        date:
          type: string
          format: date
        haircut:
          $ref: '#/components/schemas/HaircutReq'
        payment:
          $ref: '#/components/schemas/PaymentReq'
        status:
          type: string
          example: Esperando
          enum:
            - CANCELED
            - CUTTING
            - FINISHED
            - WAITING
    AppointmentRequest:
      title: AppointmentRequest
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        barber:
          $ref: '#/components/schemas/BarberReq'
        customer:
          $ref: '#/components/schemas/CustomerReq'
        date:
          type: string
          format: date
        haircut:
          $ref: '#/components/schemas/HaircutReq'
        payment:
          $ref: '#/components/schemas/PaymentReq'
        status:
          type: string
          example: Esperando
          enum:
            - CANCELED
            - CUTTING
            - FINISHED
            - WAITING
    AppointmentRes:
      title: AppointmentRes
      type: object
      properties:
        barber:
          $ref: '#/components/schemas/BarberRes'
        customer:
          $ref: '#/components/schemas/CustomerRes'
        date:
          type: string
          format: date
        haircut:
          $ref: '#/components/schemas/Haircut'
        id:
          type: integer
          format: int64
          example: 1
        payment:
          $ref: '#/components/schemas/Payment'
        status:
          type: string
          example: Esperando
...
    BarberReq:
      title: BarberReq
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
...
    BarberRes:
      title: BarberRes
      type: object
      properties:
        address:
          type: string
          example: Rua do Zé, 123
        age:
          type: integer
          format: int32
          example: 32
        email:
          type: string
          example: zebarbeiro@gmail.com
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Zé
        username:
          type: string
          example: zebarbeiro
...
    CustomerReq:
      title: CustomerReq
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
...
    CustomerRes:
      title: CustomerRes
      type: object
      properties:
        address:
          type: string
          example: Rua do Joao, 12
        age:
          type: integer
          format: int32
          example: 23
        birthDay:
          type: string
          example: 01/02/1999
        email:
          type: string
          example: joao99silva@gmail.com
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: João
        username:
          type: string
          example: joaozinho99
...
    EmptyBody:
      type: object
      nullable: true
    Haircut:
      title: Haircut
      type: object
      properties:
        description:
          type: string
          example: Barba
        id:
          type: integer
          format: int64
          example: 1
        price:
          type: number
          format: bigdecimal
          example: 9.99
    HaircutReq:
      title: HaircutReq
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
...
    Payment:
      title: Payment
      type: object
      properties:
        description:
          type: string
          example: Cartão
        id:
          type: integer
          format: int64
          example: 1
    PaymentReq:
      title: PaymentReq
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
...